feat: Rework most of component structure, use event bus system
