+++ assets/css/style.css
@@ -20,9 +20,17 @@
     border-radius: 9999px;
 }
 
+html {
+    overflow-y: hidden;
+}
+
 .character-container {
-    display: flex;
-    flex-direction: column;
+    display: grid;
+    grid-template-columns: minmax(0, 1fr) 0;
+
+    &.openMenu {
+        grid-template-columns: minmax(0, 1fr) 332px;
+    }
 }
 
 .control-grid {
@@ -31,11 +39,6 @@
 }
 
 @media (min-width: 1536px) {
-    .character-container {
-        display: grid;
-        grid-template-columns: minmax(0, 1fr) 332px;
-    }
-
     .control-grid {
         grid-template-columns: 290px minmax(0, 1fr) 290px;
         grid-template-rows: 1fr;
