+++ models/GetCharactersRequest.ts
@@ -4,10 +4,10 @@ export default class GetCharactersRequest {
     public readonly Order: string;
     public readonly Filter: string;
 
-    public constructor(page: number, count: number, order: string, filter: string) {
+    public constructor(page: number, filter: string, count: number, order: string) {
         this.Page = page;
+        this.Filter = filter;
         this.Count = count;
         this.Order = order;
-        this.Filter = filter;
     }
 }
