+++ stores/controlComponentStore.ts
@@ -1,0 +1,10 @@
+import type { FileUpload } from '~/models/OLD/FileUpload';
+
+export const useControlComponentStore = defineStore('controlComponent', {
+    state: () => {
+        return {
+            processing: false,
+            files: [] as FileUpload[],
+        };
+    },
+});
