+++ utils/sleep.ts
@@ -1,0 +1,3 @@
+export default function sleep(milliseconds: number) {
+    return new Promise((resolve) => setTimeout(resolve, milliseconds));
+}
